feat: rename folder "types" to "database" so that module name in generated by typedoc make more sense
