+++ bindings/javascript/src/database/common.ts
