+++ bindings/javascript/src/database/database.ts
