+++ bindings/javascript/src/database/entities/circle.ts
