+++ bindings/javascript/src/database/entities/dimension.ts
