+++ bindings/javascript/src/database/entities/ellipse.ts
