+++ bindings/javascript/src/database/entities/entity.ts
