+++ bindings/javascript/src/database/entities/image.ts
