+++ bindings/javascript/src/database/entities/insert.ts
