+++ bindings/javascript/src/database/entities/lwpolyline.ts
