+++ bindings/javascript/src/database/entities/mtext.ts
