+++ bindings/javascript/src/database/entities/point.ts
