+++ bindings/javascript/src/database/entities/polyline.ts
