+++ bindings/javascript/src/database/entities/ray.ts
