+++ bindings/javascript/src/database/entities/section.ts
