+++ bindings/javascript/src/database/entities/spline.ts
