+++ bindings/javascript/src/database/entities/table.ts
