+++ bindings/javascript/src/database/entities/text.ts
