+++ bindings/javascript/src/database/entities/vertex.ts
