+++ bindings/javascript/src/database/objects/common.ts
