+++ bindings/javascript/src/database/objects/imageDef.ts
