+++ bindings/javascript/src/database/objects/index.ts
