+++ bindings/javascript/src/database/shared/index.ts
