+++ bindings/javascript/src/database/shared/xdata.ts
