+++ bindings/javascript/src/database/tables/blockRecord.ts
