+++ bindings/javascript/src/database/tables/dimStyle.ts
