+++ bindings/javascript/src/database/tables/index.ts
