+++ bindings/javascript/src/database/tables/layer.ts
