+++ bindings/javascript/src/database/tables/ltype.ts
