+++ bindings/javascript/src/database/tables/style.ts
