+++ bindings/javascript/src/database/tables/table.ts
